Fixed some memory function issues

+++ header.h
@@ -84,8 +84,7 @@ int readData(void);
 uint16_t string_to_octal(string input_string);
 string octal_to_string(uint16_t value);
 uint16_t read_byte(uint16_t address, bool trace = true);                              //Read a byte and return it in the low 8 bits
-//uint16_t read_word(uint16_t address, bool trace = true, bool is_instruction = false); //Read two bytes in memory and return as little endian word
-uint16_t read_word(int address, bool trace = true, bool is_instruction = false); //Read two bytes in memory and return as little endian word
+uint16_t read_word(uint16_t address, bool trace = true, bool is_instruction = false); //Read two bytes in memory and return as little endian word
 void write_byte(uint16_t address, uint16_t byte, bool trace = true);                  //Take byte in the low 8 bits and write to memory
 void write_word(uint16_t address, uint16_t word, bool trace = true);                  //Write little endian word to two bytes in memory
 void print_octal(uint16_t value);
